Rename config.js to index.js
