+++ src/annotator/config/index.js
